Updates to s_max
The estimate previously used for s_max was conservative in the sense that it
assumed perfect/spectral derivatives. Since all of our discrete Laplacians
are second-order, I've updated it with a less conservative estimate based
on second-order differencing. The result is that it requires fewer stages
to reach a desired level of accuracy. I also cleaned up some code comments.
tests will be updated in a separate commit.