Disentangled from _compute_filter_spec
This commit puts the things needed to plot the filter approximation into FilterSpec. That way plot_shape can be a separate method that leaves _compute_filter_spec
unchanged, except for the fact that s_max and p are now part of the output of _compute_filter_spec.

I had to update the filter spec tests by defining the target/true values to have values of s_max and p. We don't test these values, so I just set them to 0.