Option to Factor Gaussian
This commit introduces the variable `n_iterations`. When it's 1,
nothing changes. When it's > 1, the Gaussian filter is implemented
by applying a sequence of `n_iterations` identical Gaussian filters
each of which has a length scale `filter_scale / np.sqrt(n_iterations)`.
The idea is to avoid numerical instability for very large filter
factors by splitting the filter up into a sequence of smaller
filters.